(bluefox) support of compact mode (bluefox) add error codes

+++ controller.js
@@ -93,6 +93,7 @@ let inputCount              = 0;
 let outputCount             = 0;
 let mhService               = null; // multihost service
 const uptimeStart             = Date.now();
+const adapterModules          = {};
 
 const config = getConfig();
 
@@ -1315,9 +1316,9 @@ function getVersionFromHost(hostId, callback) {
     });
 }
 /**
-    Helper function that serialize deletion of states
-    @param {object} list array with states
-    @param {function} cb optional callback
+ Helper function that serialize deletion of states
+ @param {object} list array with states
+ @param {function} cb optional callback
  */
 function _deleteAllZipPackages(list, cb) {
     if (!list || !list.length) {
@@ -1329,10 +1330,10 @@ function _deleteAllZipPackages(list, cb) {
     }
 }
 /**
-    This function deletes all ZIP packages that were not downloaded.
-    ZIP Package is temporary file, that should be deleted straight after it downloaded and if it still exists, so clear it
+ This function deletes all ZIP packages that were not downloaded.
+ ZIP Package is temporary file, that should be deleted straight after it downloaded and if it still exists, so clear it
 
-    @param {function} cb optional callback
+ @param {function} cb optional callback
  */
 function deleteAllZipPackages(cb) {
     states.getKeys('system.host.' + hostname + '.zip.*',
@@ -1636,7 +1637,7 @@ function processMessage(msg) {
                             lines.push(stats.size);
                             sendTo(msg.from, msg.command, lines, msg.callback);
                         }).on('error', () =>   // done
-                            sendTo(msg.from, msg.command, [stats.size], msg.callback));
+                        sendTo(msg.from, msg.command, [stats.size], msg.callback));
                 } else {
                     sendTo(msg.from, msg.command, [0], msg.callback);
                 }
@@ -2035,7 +2036,7 @@ function storePids() {
             for (const id in procs) {
                 if (!procs.hasOwnProperty(id)) continue;
 
-                if (procs[id].process) {
+                if (procs[id].process && procs[id].process.pid) {
                     pids.push(procs[id].process.pid);
                 }
                 pids.push(process.pid);
@@ -2239,17 +2240,41 @@ function startInstance(id, wakeUp) {
         }
     }
 
+    procs[id].startedInCompactMode = false;
+
     switch (mode) {
         case 'once':
         case 'daemon':
             if (procs[id] && !procs[id].process) {
                 allInstancesStopped = false;
                 logger.debug('host.' + hostname + ' startInstance ' + name + '.' + args[0] + ' loglevel=' + args[1]);
-                procs[id].process = cp.fork(fileNameFull, args, {stdio: ['ignore', 'ignore', 'pipe', 'ipc']});
+                if (config.system.compact && instance.common.compact) {
+                    if (!adapterModules[name] && fileNameFull) {
+                        try {
+                            adapterModules[name] = require(fileNameFull);
+                        } catch (e) {
+                            logger.error(`host.${hostname} error with ${fileNameFull}: ${JSON.stringify(e)}`);
+                        }
+                    }
+                    if (!adapterModules[name]) {
+                        procs[id].process = cp.fork(fileNameFull, args, {stdio: ['ignore', 'ignore', 'pipe', 'ipc']});
+                    } else {
+                        const _instance = (instance && instance._id && instance.common) ? instance._id.split('.').pop() || 0 : 0;
+                        const logLevel = (instance && instance._id && instance.common) ? instance.common.loglevel || 'info' : 'info';
+                        try {
+                            procs[id].process = adapterModules[name]({logLevel, compactInstance: _instance, compact: true});
+                            procs[id].startedInCompactMode = true;
+                        } catch (e) {
+                            logger.error(`host.${hostname} Cannot start ${name}.${_instance}: ${JSON.stringify(e)}`);
+                        }
+                    }
+                } else {
+                    procs[id].process = cp.fork(fileNameFull, args, {stdio: ['ignore', 'ignore', 'pipe', 'ipc']});
+                }
 
                 states.setState(id + '.sigKill', {val: procs[id].process.pid, ack: true, from: 'system.host.' + hostname});
                 // catch error output
-                if (procs[id].process.stderr) {
+                if (!procs[id].startedInCompactMode && procs[id].process.stderr) {
                     procs[id].process.stderr.on('data', data => {
                         if (!data || !procs[id] || typeof procs[id] !== 'object') return;
                         const text = data.toString();
@@ -2462,11 +2487,21 @@ function stopInstance(id, callback) {
 
     const instance = procs[id].config;
     if (!instance || !instance.common || !instance.common.mode) {
-        if (procs[id].process) {
+        if (procs[id].process && !procs[id].startedInCompactMode) {
             procs[id].stopping = true;
             procs[id].process.kill();
             delete procs[id].process;
         }
+        if (procs[id].startedInCompactMode) {
+            procs[id].stopping = true;
+
+            try {
+                procs[id].process.stop();  // call stop directly in adapter.js
+            } catch (e) {
+                logger.error(`host.${hostname} Cannot stop ${id}: ${JSON.stringify(e)}`);
+            }
+        }
+
         if (procs[id].schedule) {
             procs[id].schedule.cancel();
             delete procs[id].schedule;
